Dvision CastAndroid.java Thread module to ActionRunnable, and added new pthon command handling, added new build of *.jar plugin
